+++ site-lisp/emacs-doom-themes/themes/doom-fairy-floss-theme.el
@@ -1,4 +1,4 @@
-;;; doom-fairy-floss-theme.el --- inspired by sailorhg Fairy Floss
+;;; doom-fairy-floss-theme.el --- inspired by sailorhg Fairy Floss -*- no-byte-compile: t; -*-
 (require 'doom-themes)
 
 ;;
@@ -120,6 +120,9 @@ determine the exact padding."
    (font-lock-doc-face
     :inherit 'font-lock-comment-face
     :foreground doc-comments)
+   (font-lock-keyword-face
+    :slant 'italic
+    :foreground keywords)
 
    ;; Centaur tabs
    (centaur-tabs-active-bar-face :background blue)