+++ site-lisp/emacs-doom-themes/themes/doom-vibrant-theme.el
@@ -1,4 +1,4 @@
-;; doom-vibrant-theme.el --- a more vibrant version of doom-one
+;; doom-vibrant-theme.el --- a more vibrant version of doom-one -*- no-byte-compile: t; -*-
 (require 'doom-themes)
 
 ;;
@@ -108,7 +108,8 @@ determine the exact padding."
 
 
   ;; --- extra faces ------------------------
-  ((elscreen-tab-other-screen-face :background "#353a42" :foreground "#1e2022")
+  ((centaur-tabs-unselected :background bg-alt :foreground base6)
+   (elscreen-tab-other-screen-face :background "#353a42" :foreground "#1e2022")
 
    (font-lock-comment-face
     :foreground comments