+++ site-lisp/lsp-mode/lsp-rust.el
@@ -36,9 +36,9 @@
 
 (defcustom lsp-rust-server 'rls
   "Choose LSP server."
-  :type '(choice (symbol :tag 'rls "rls")
-                 (symbol :tag 'rust-analyzer "rust-analyzer"))
-  :group 'lsp-mode
+  :type '(choice (const :tag "rls" rls)
+                 (const :tag "rust-analyzer" rust-analyzer))
+  :group 'lsp-rust
   :package-version '(lsp-mode . "6.2"))
 
 ;; RLS
@@ -335,6 +335,74 @@ PARAMS progress report notification data."
   :type 'boolean
   :package-version '(lsp-mode . "6.2"))
 
+(defcustom lsp-rust-analyzer-max-inlay-hint-length nil
+  "Max inlay hint length."
+  :type 'integer
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-lru-capacity nil
+  "LRU capacity."
+  :type 'integer
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-cargo-watch-enable t
+  "Enable Cargo watch."
+  :type 'boolean
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-cargo-watch-command "check"
+  "Cargo watch command."
+  :type 'string
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-cargo-watch-args []
+  "Cargo watch args."
+  :type 'lsp-string-vector
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-cargo-all-targets nil
+  "Cargo watch all targets or not."
+  :type 'boolean
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-use-client-watching t
+  "Use client watching"
+  :type 'boolean
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-exclude-globs []
+  "Exclude globs"
+  :type 'lsp-string-vector
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-enabled-feature-flags ["completion.insertion.add-call-parenthesis"
+                                                    "completion.enable-postfix"
+                                                    "notifications.workspace-loaded"]
+  "Feature flags to set."
+  :type 'lsp-string-vector
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defcustom lsp-rust-analyzer-macro-expansion-method 'lsp-rust-analyzer-macro-expansion-default
+  "Use a different function if you want formatted macro expansion results and syntax highlighting."
+  :type 'function
+  :package-version '(lsp-mode . "6.2.2"))
+
+(defun lsp-rust-analyzer--make-init-options ()
+  "Init options for rust-analyzer"
+  (let ((feature-flags (--map (cons (intern it) t) lsp-rust-analyzer-enabled-feature-flags)))
+    `(:lruCapacity ,lsp-rust-analyzer-lru-capacity
+      :maxInlayHintLength ,lsp-rust-analyzer-max-inlay-hint-length
+      :cargoWatchEnable ,(lsp-json-bool lsp-rust-analyzer-cargo-watch-enable)
+      :cargoWatchCommand ,lsp-rust-analyzer-cargo-watch-command
+      :cargoWatchArgs ,lsp-rust-analyzer-cargo-watch-args
+      :cargoWatchAllTargets ,(lsp-json-bool lsp-rust-analyzer-cargo-all-targets)
+      :excludeGlobs ,lsp-rust-analyzer-exclude-globs
+      :useClientWatching ,(lsp-json-bool lsp-rust-analyzer-use-client-watching)
+      :featureFlags ,feature-flags
+      :cargoFeatures (:allFeatures ,(lsp-json-bool lsp-rust-all-features)
+                      :noDefaultFeatures ,(lsp-json-bool lsp-rust-no-default-features)
+                      :features ,lsp-rust-features))))
+
 (defconst lsp-rust-notification-handlers
   '(("rust-analyzer/publishDecorations" . (lambda (_w _p)))))
 
@@ -419,8 +487,10 @@ PARAMS progress report notification data."
   :new-connection (lsp-stdio-connection (lambda () lsp-rust-analyzer-server-command))
   :major-modes '(rust-mode rustic-mode)
   :priority (if (eq lsp-rust-server 'rust-analyzer) 1 -1)
+  :initialization-options 'lsp-rust-analyzer--make-init-options
   :notification-handlers (ht<-alist lsp-rust-notification-handlers)
   :action-handlers (ht<-alist lsp-rust-action-handlers)
+  :library-folders-fn (lambda (_workspace) lsp-rust-library-directories)
   :ignore-messages nil
   :server-id 'rust-analyzer))
 
@@ -483,6 +553,32 @@ PARAMS progress report notification data."
   (add-hook 'lsp-after-open-hook (lambda ()
                                    (when (lsp-find-workspace 'rust-analyzer nil)
                                      (lsp-rust-analyzer-inlay-hints-mode)))))
+
+(defun lsp-rust-analyzer-expand-macro ()
+  "Expands the macro call at point recursively."
+  (interactive)
+  (-if-let (workspace (lsp-find-workspace 'rust-analyzer default-directory))
+      (-if-let* ((params (list :textDocument (lsp--text-document-identifier)
+                               :position (lsp--cur-position)))
+                 (response (with-lsp-workspace workspace
+                             (lsp-send-request (lsp-make-request
+                                                "rust-analyzer/expandMacro"
+                                                params))))
+                 (result (ht-get response "expansion")))
+          (funcall lsp-rust-analyzer-macro-expansion-method result)
+        (message "No macro found at point, or it could not be expanded."))
+    (message "rust-analyzer not running.")))
+
+(defun lsp-rust-analyzer-macro-expansion-default (result)
+  "Default method for displaying macro expansion."
+  (let* ((root (lsp-workspace-root default-directory))
+         (buf (get-buffer-create (get-buffer-create (format "*rust-analyzer macro expansion %s*" root)))))
+    (with-current-buffer buf
+      (let ((inhibit-read-only t))
+        (erase-buffer)
+        (insert result)
+        (special-mode)))
+    (display-buffer buf)))
 
 (provide 'lsp-rust)
 ;;; lsp-rust.el ends here